requirements.txt
pyyaml
aiofiles
fastapi
starlette
pydantic
opencv-python-headless
numpy
chromadb
ultralytics
Pillow
datetime
uvicorn
passlib==1.7.4
python-multipart
pydantic[email]
qdrant-client
psycopg2-binary
websockets
httpx
langchain 
langchain-ollama 
#jwt
python-dotenv
xgboost
anyio
#smtplib
#aiosmtplib
#sqlite3
asyncpg
pyotp
langchain_community
PyJWT
tenacity
apscheduler

pytest-mock
pytest-asyncio
pytest-tornasync
pytest-trio
pytest-twisted
pytest 
pytest-asyncio
pytest-cov
pandas
elasticsearch
argon2-cffi